Moving internal to pkg (so brave people could import the schemaconverter package for their own use)
