+++ Makefile
@@ -16,12 +16,12 @@ build_linux:
 samples: build
 	@echo "Generating sample JSON-Schemas ..."
 	@mkdir -p out
-	@bin/openapi2jsonschema  -spec=internal/schemaconverter/samples/swagger2/flat-object.yaml                 -go_constants  -block_additional_properties  -out=./out
-	@bin/openapi2jsonschema  -spec=internal/schemaconverter/samples/swagger2/referenced-object.yaml           -go_constants  -block_additional_properties  -out=./out
-	@bin/openapi2jsonschema  -spec=internal/schemaconverter/samples/swagger2/object-with-pattern.yaml         -go_constants  -block_additional_properties  -out=./out
-	@bin/openapi2jsonschema  -spec=internal/schemaconverter/samples/swagger2/array-of-referenced-object.yaml  -go_constants  -block_additional_properties  -out=./out
-	@bin/openapi2jsonschema  -spec=internal/schemaconverter/samples/openapi3/with_map.yaml                    -go_constants  -block_additional_properties  -out=./out -v3
-	@bin/openapi2jsonschema  -spec=internal/schemaconverter/samples/openapi3/petstore.yaml                    -go_constants  -block_additional_properties  -out=./out -v3
+	@bin/openapi2jsonschema  -spec=pkg/schemaconverter/samples/swagger2/flat-object.yaml                 -go_constants  -block_additional_properties  -out=./out
+	@bin/openapi2jsonschema  -spec=pkg/schemaconverter/samples/swagger2/referenced-object.yaml           -go_constants  -block_additional_properties  -out=./out
+	@bin/openapi2jsonschema  -spec=pkg/schemaconverter/samples/swagger2/object-with-pattern.yaml         -go_constants  -block_additional_properties  -out=./out
+	@bin/openapi2jsonschema  -spec=pkg/schemaconverter/samples/swagger2/array-of-referenced-object.yaml  -go_constants  -block_additional_properties  -out=./out
+	@bin/openapi2jsonschema  -spec=pkg/schemaconverter/samples/openapi3/with_map.yaml                    -go_constants  -block_additional_properties  -out=./out -v3
+	@bin/openapi2jsonschema  -spec=pkg/schemaconverter/samples/openapi3/petstore.yaml                    -go_constants  -block_additional_properties  -out=./out -v3
 
 test:
 	@go test ./... -cover
