+++ pkg/schemaconverter/oapi2/jsonschema_allow_nulls_test.go
@@ -1,10 +1,10 @@
-package oapi3
+package oapi2
 
 import (
 	"fmt"
 	"testing"
 
-	"github.com/chrusty/openapi2jsonschema/internal/schemaconverter/types"
+	"github.com/chrusty/openapi2jsonschema/pkg/schemaconverter/types"
 	"github.com/sirupsen/logrus"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -61,7 +61,7 @@ func TestGenerateJSONSchemasAllowNullsFlatObject(t *testing.T) {
 		AllowNullValues:           true,
 		BlockAdditionalProperties: false,
 		JSONSchemaFileExtention:   "jsonschema",
-		SpecPath:                  "../samples/openapi3/flat-object.yaml",
+		SpecPath:                  "../samples/swagger2/flat-object.yaml",
 	}, logrus.New())
 	require.NoError(t, err)
 
@@ -117,7 +117,7 @@ func TestGenerateJSONSchemasAllowNullsFlatObjectWithEnum(t *testing.T) {
 		AllowNullValues:           true,
 		BlockAdditionalProperties: false,
 		JSONSchemaFileExtention:   "jsonschema",
-		SpecPath:                  "../samples/openapi3/flat-object-with-enum.yaml",
+		SpecPath:                  "../samples/swagger2/flat-object-with-enum.yaml",
 	}, logrus.New())
 	require.NoError(t, err)
 
@@ -312,7 +312,7 @@ func TestGenerateJSONSchemasAllowNullsObjectWithArrays(t *testing.T) {
 		AllowNullValues:           true,
 		BlockAdditionalProperties: false,
 		JSONSchemaFileExtention:   "jsonschema",
-		SpecPath:                  "../samples/openapi3/array-of-referenced-object.yaml",
+		SpecPath:                  "../samples/swagger2/array-of-referenced-object.yaml",
 	}, logrus.New())
 	require.NoError(t, err)
 
@@ -402,7 +402,7 @@ func TestGenerateJSONSchemasAllowNullsObjectWithPattern(t *testing.T) {
 		AllowNullValues:           true,
 		BlockAdditionalProperties: false,
 		JSONSchemaFileExtention:   "jsonschema",
-		SpecPath:                  "../samples/openapi3/object-with-pattern.yaml",
+		SpecPath:                  "../samples/swagger2/object-with-pattern.yaml",
 	}, logrus.New())
 	require.NoError(t, err)
 
@@ -620,7 +620,7 @@ func TestGenerateJSONSchemasAllowNullsReferencedObject(t *testing.T) {
 		AllowNullValues:           true,
 		BlockAdditionalProperties: false,
 		JSONSchemaFileExtention:   "jsonschema",
-		SpecPath:                  "../samples/openapi3/referenced-object.yaml",
+		SpecPath:                  "../samples/swagger2/referenced-object.yaml",
 	}, logrus.New())
 	require.NoError(t, err)
 
@@ -690,7 +690,7 @@ func TestGenerateJSONSchemasAllowNullsNumberWithMinMax(t *testing.T) {
 		AllowNullValues:           true,
 		BlockAdditionalProperties: false,
 		JSONSchemaFileExtention:   "jsonschema",
-		SpecPath:                  "../samples/openapi3/flat-object-with-number-options.yaml",
+		SpecPath:                  "../samples/swagger2/flat-object-with-number-options.yaml",
 	}, logrus.New())
 	require.NoError(t, err)
 
@@ -701,9 +701,7 @@ func TestGenerateJSONSchemasAllowNullsNumberWithMinMax(t *testing.T) {
 	assert.NoError(t, err)
 	assert.NotNil(t, generatedJSONSchemas)
 	assert.Len(t, generatedJSONSchemas, 1)
-	if !assert.JSONEq(t, expectedSchema, string(generatedJSONSchemas[0].Bytes)) {
-		fmt.Println(string(generatedJSONSchemas[0].Bytes))
-	}
+	assert.JSONEq(t, expectedSchema, string(generatedJSONSchemas[0].Bytes))
 }
 
 func TestGenerateJSONSchemasMapAllowingNullValues(t *testing.T) {
@@ -736,7 +734,7 @@ func TestGenerateJSONSchemasMapAllowingNullValues(t *testing.T) {
 		AllowNullValues:           true,
 		BlockAdditionalProperties: false,
 		JSONSchemaFileExtention:   "jsonschema",
-		SpecPath:                  "../samples/openapi3/with_map.yaml",
+		SpecPath:                  "../samples/swagger2/with_map.yaml",
 	}, logrus.New())
 	require.NoError(t, err)
 
@@ -795,7 +793,7 @@ func TestGenerateJSONSchemasMapInAReffedObjectAllowingNullValues(t *testing.T) {
 		AllowNullValues:           true,
 		BlockAdditionalProperties: false,
 		JSONSchemaFileExtention:   "jsonschema",
-		SpecPath:                  "../samples/openapi3/with_map_in_ref.yaml",
+		SpecPath:                  "../samples/swagger2/with_map_in_ref.yaml",
 	}, logrus.New())
 	require.NoError(t, err)
 
@@ -809,6 +807,7 @@ func TestGenerateJSONSchemasMapInAReffedObjectAllowingNullValues(t *testing.T) {
 	if !assert.JSONEq(t, expectedSchema, string(generatedJSONSchemas[1].Bytes)) {
 		fmt.Println(string(generatedJSONSchemas[1].Bytes))
 	}
+
 }
 
 func TestGenerateJSONSchemasMapInAReffedObjectAllowingNullValues2(t *testing.T) {
@@ -846,7 +845,7 @@ func TestGenerateJSONSchemasMapInAReffedObjectAllowingNullValues2(t *testing.T) 
 		AllowNullValues:           true,
 		BlockAdditionalProperties: false,
 		JSONSchemaFileExtention:   "jsonschema",
-		SpecPath:                  "../samples/openapi3/with_map_in_ref_2.yaml",
+		SpecPath:                  "../samples/swagger2/with_map_in_ref_2.yaml",
 	}, logrus.New())
 	require.NoError(t, err)
