+++ pkg/schemaconverter/oapi3/mapping.go
@@ -1,16 +1,14 @@
-package oapi2
+package oapi3
 
 import (
 	"encoding/json"
 	"fmt"
 	"sort"
-	"strconv"
 	"strings"
 
-	"github.com/chrusty/openapi2jsonschema/internal/schemaconverter/types"
-
-	openAPI "github.com/NYTimes/openapi2proto/openapi"
 	jsonSchema "github.com/alecthomas/jsonschema"
+	"github.com/chrusty/openapi2jsonschema/pkg/schemaconverter/types"
+	"github.com/getkin/kin-openapi/openapi3"
 	"github.com/pkg/errors"
 	"github.com/xeipuuv/gojsonschema"
 )
@@ -19,20 +17,8 @@ import (
 func (c *Converter) mapOpenAPIDefinitionsToJSONSchema() ([]types.GeneratedJSONSchema, error) {
 	var generatedJSONSchemas []types.GeneratedJSONSchema
 
-	// // if we have no definitions then copy them from parameters:
-	// if c.spec.Definitions == nil {
-	// 	c.logger.Debug("No definitions found - copying from parameters")
-	// 	c.spec.Definitions = map[string]*openAPI.Schema{}
-
-	// 	// jam all the parameters into the normal 'definitions' for easier reference.
-	// 	for paramName, param := range c.spec.Parameters {
-	// 		c.logger.WithField("parameter_name", paramName).Trace("Found a parameter")
-	// 		c.spec.Parameters[paramName] = param
-	// 	}
-	// }
-
 	// Iterate through any schemas we find, creating JSONSchemas for each:
-	for schemaName, schema := range c.spec.Definitions {
+	for schemaName, schema := range c.swagger.Components.Schemas {
 		var generatedJSONSchema types.GeneratedJSONSchema
 
 		c.logger.WithField("schema_name", schemaName).Trace("Found a schema")
@@ -61,23 +47,32 @@ func (c *Converter) mapOpenAPIDefinitionsToJSONSchema() ([]types.GeneratedJSONSc
 }
 
 // convertSchema converts an OpenAPI Schema into a JSON-Schema:
-func (c *Converter) convertSchema(openAPISchema *openAPI.Schema) (jsonSchema.Type, error) {
+func (c *Converter) convertSchema(openAPISchema *openapi3.SchemaRef) (jsonSchema.Type, error) {
 
 	// Prepare a new jsonschema:
 	definitionJSONSchema := jsonSchema.Type{
 		AdditionalProperties: c.generateAdditionalProperties(),
-		Description:          strings.Replace(openAPISchema.Description, "`", "'", -1),
-		MaxLength:            openAPISchema.MaxLength,
-		MinLength:            openAPISchema.MinLength,
-		Pattern:              openAPISchema.Pattern,
+		Description:          strings.Replace(openAPISchema.Value.Description, "`", "'", -1),
+		MinLength:            int(openAPISchema.Value.MinLength),
+		Pattern:              openAPISchema.Value.Pattern,
 		Properties:           make(map[string]*jsonSchema.Type),
-		Minimum:              openAPISchema.Minimum,
-		Maximum:              openAPISchema.Maximum,
+	}
+
+	if openAPISchema.Value.MaxLength != nil {
+		definitionJSONSchema.MaxLength = int(*openAPISchema.Value.MaxLength)
+	}
+
+	if openAPISchema.Value.Min != nil {
+		definitionJSONSchema.Minimum = int(*openAPISchema.Value.Min)
+	}
+
+	if openAPISchema.Value.Max != nil {
+		definitionJSONSchema.Maximum = int(*openAPISchema.Value.Max)
 	}
 
 	// Arrays of self-defined parameters:
-	if openAPISchema.Ref == "" && openAPISchema.Type.Contains(gojsonschema.TYPE_ARRAY) {
-		itemsMap, err := c.recurseNestedSchemas(map[string]*openAPI.Schema{"items": openAPISchema.Items})
+	if openAPISchema.Ref == "" && strings.Contains(openAPISchema.Value.Type, gojsonschema.TYPE_ARRAY) {
+		itemsMap, err := c.recurseNestedSchemas(map[string]*openapi3.SchemaRef{"items": openAPISchema.Value.Items})
 		if err != nil {
 			return definitionJSONSchema, err
 		}
@@ -85,19 +80,19 @@ func (c *Converter) convertSchema(openAPISchema *openAPI.Schema) (jsonSchema.Typ
 	}
 
 	// Single-instances of self-defined parameters:
-	if openAPISchema.Ref == "" && !openAPISchema.Type.Contains(gojsonschema.TYPE_ARRAY) && openAPISchema.Items == nil {
-		properties, err := c.recurseNestedSchemas(openAPISchema.Properties)
+	if openAPISchema.Ref == "" && !strings.Contains(openAPISchema.Value.Type, gojsonschema.TYPE_ARRAY) && openAPISchema.Value.Items == nil {
+		properties, err := c.recurseNestedSchemas(openAPISchema.Value.Properties)
 		definitionJSONSchema.Properties = properties
 		if err != nil {
 			return definitionJSONSchema, err
 		}
 
 		// See if there are any additionalProperties to convert:
-		if openAPISchema.AdditionalProperties != nil {
-			if convertedAdditionalProperties, err := c.convertSchema(openAPISchema.AdditionalProperties); err == nil {
+		if openAPISchema.Value.AdditionalProperties != nil {
+			if convertedAdditionalProperties, err := c.convertSchema(openAPISchema.Value.AdditionalProperties); err == nil {
 				c.logger.
-					WithField("AdditionalProperties.Ref", openAPISchema.AdditionalProperties.Ref).
-					WithField("AdditionalProperties.Type", openAPISchema.AdditionalProperties.Type).
+					WithField("AdditionalProperties.Value.Ref", openAPISchema.Value.AdditionalProperties.Ref).
+					WithField("AdditionalProperties.Value.Type", openAPISchema.Value.AdditionalProperties.Value.Type).
 					Tracef("Converted additional properties: %v", convertedAdditionalProperties)
 				additionalPropertiesJSON, err := json.Marshal(convertedAdditionalProperties)
 				if err != nil {
@@ -111,23 +106,22 @@ func (c *Converter) convertSchema(openAPISchema *openAPI.Schema) (jsonSchema.Typ
 		if c.config.AllowNullValues {
 			definitionJSONSchema.OneOf = []*jsonSchema.Type{
 				{Type: gojsonschema.TYPE_NULL},
-				{Type: c.mapOpenAPITypeToJSONSchemaType(openAPISchema.Type)},
+				{Type: c.mapOpenAPITypeToJSONSchemaType(openAPISchema.Value.Type)},
 			}
 		} else {
-			definitionJSONSchema.Type = c.mapOpenAPITypeToJSONSchemaType(openAPISchema.Type)
+			definitionJSONSchema.Type = c.mapOpenAPITypeToJSONSchemaType(openAPISchema.Value.Type)
 		}
 
-		definitionJSONSchema.Required = openAPISchema.Required
-		definitionJSONSchema.Enum = c.mapEnums(openAPISchema.Enum, openAPISchema.Type)
+		definitionJSONSchema.Required = openAPISchema.Value.Required
+		definitionJSONSchema.Enum = openAPISchema.Value.Enum
 
-		if openAPISchema.Format != "" {
-			definitionJSONSchema.Format = openAPISchema.Format
+		if openAPISchema.Value.Format != "" {
+			definitionJSONSchema.Format = openAPISchema.Value.Format
 		}
 	}
 
 	// Referenced models:
 	if openAPISchema.Ref != "" {
-		var enum []string
 		var lookedupReferenceType string
 		nestedProperties, lookedupReferenceType, required, enum, err := c.lookupReference(openAPISchema.Ref)
 		if err != nil {
@@ -143,7 +137,7 @@ func (c *Converter) convertSchema(openAPISchema *openAPI.Schema) (jsonSchema.Typ
 			definitionJSONSchema.Type = lookedupReferenceType
 		}
 		definitionJSONSchema.Properties, err = c.recurseNestedSchemas(nestedProperties)
-		definitionJSONSchema.Enum = c.mapEnums(enum, []string{definitionJSONSchema.Type})
+		definitionJSONSchema.Enum = enum
 
 		if openAPISchema.Ref != "" {
 			referenceName, _ := c.splitReferencePath(openAPISchema.Ref)
@@ -158,9 +152,12 @@ func (c *Converter) convertSchema(openAPISchema *openAPI.Schema) (jsonSchema.Typ
 	// Maintain a list of required items:
 	if definitionJSONSchema.Type == gojsonschema.TYPE_OBJECT {
 
-		// if we have any nested items in the object then we should process them
-		if additionalPropertiesSchema := openAPISchema.AdditionalProperties; additionalPropertiesSchema != nil {
-			schema, err := c.convertSchema(additionalPropertiesSchema)
+		// If we have any nested items in the object then we should process them:
+		if openAPISchema.Value.AdditionalProperties != nil {
+			schema, err := c.convertSchema(openAPISchema.Value.AdditionalProperties)
+			if err != nil {
+				return definitionJSONSchema, err
+			}
 
 			// Annoyingly since "additionalProperties" can actually be a
 			// boolean or an object we have to marshal the resulting schema
@@ -174,34 +171,17 @@ func (c *Converter) convertSchema(openAPISchema *openAPI.Schema) (jsonSchema.Typ
 	return definitionJSONSchema, nil
 }
 
-// mapEnums maps OpenAPI enums to JSONSchema types:
-func (c *Converter) mapEnums(items []string, openAPISchemaTypes openAPI.SchemaType) []interface{} {
-	var result []interface{}
-
-	for _, item := range items {
-		var value interface{}
-		if openAPISchemaTypes.Contains(gojsonschema.TYPE_NUMBER) {
-			value, _ = strconv.Atoi(item)
-		} else {
-			value = item
-		}
-		result = append(result, value)
-	}
-
-	return result
-}
-
 // mapOpenAPITypeToJSONSchemaType maps OpenAPI types to JSONSchema types:
-func (c *Converter) mapOpenAPITypeToJSONSchemaType(openAPISchemaTypes openAPI.SchemaType) string {
+func (c *Converter) mapOpenAPITypeToJSONSchemaType(openAPISchemaType string) string {
 
 	// Make sure we were actually given a type:
-	if len(openAPISchemaTypes) != 1 {
-		c.logger.WithField("type", openAPISchemaTypes).Error("Can't determine JSONSchema type")
+	if openAPISchemaType == "" {
+		c.logger.WithField("type", openAPISchemaType).Warn("Can't determine JSONSchema type")
 		return gojsonschema.TYPE_NULL
 	}
 
 	// Switch on the first type:
-	switch openAPISchemaTypes[0] {
+	switch openAPISchemaType {
 	case "array":
 		return gojsonschema.TYPE_ARRAY
 	case "boolean":
@@ -217,26 +197,26 @@ func (c *Converter) mapOpenAPITypeToJSONSchemaType(openAPISchemaTypes openAPI.Sc
 	case "":
 		return gojsonschema.TYPE_NULL
 	default:
-		c.logger.WithField("type", openAPISchemaTypes).Warn("Can't determine JSONSchema type")
+		c.logger.WithField("type", openAPISchemaType).Warn("Can't determine JSONSchema type")
 		return gojsonschema.TYPE_NULL
 	}
 }
 
-// splitReferencePath breaks up a reference path into its components (OpenAPI2 references look like "#/definitions/Something"):
+// splitReferencePath breaks up a reference path into its components (OpenAPI3 references look like "#/components/schemas/Something"):
 func (c *Converter) splitReferencePath(ref string) (string, error) {
 
 	// split on '/':
 	refDatas := strings.Split(ref, "/")
 
-	// Return the 3rd component (definition name):
-	if len(refDatas) > 1 {
-		return refDatas[2], nil
+	// Return the 4th component (definition name):
+	if len(refDatas) > 2 {
+		return refDatas[3], nil
 	}
 	return "", fmt.Errorf("Unable to split this reference (%s)", ref)
 }
 
 // lookupReference looks up a reference and returns its schema and metadata:
-func (c *Converter) lookupReference(referencePath string) (nestedProperties map[string]*openAPI.Schema, definitionJSONSchemaType string, requiredProperties []string, enum []string, err error) {
+func (c *Converter) lookupReference(referencePath string) (nestedProperties map[string]*openapi3.SchemaRef, definitionJSONSchemaType string, requiredProperties []string, enum []interface{}, err error) {
 	c.logger.WithField("referencePath", referencePath).Trace("Looking up reference")
 
 	// Break up the path:
@@ -247,18 +227,18 @@ func (c *Converter) lookupReference(referencePath string) (nestedProperties map[
 
 	// Look up the referenced model:
 	c.logger.WithField("reference", referenceName).Trace("Found a referenced model")
-	referencedDefinition, ok := c.spec.Definitions[referenceName]
+	referencedDefinition, ok := c.swagger.Components.Schemas[referenceName]
 	if !ok {
 		err = fmt.Errorf("Unable to find a referenced model (%s)", referenceName)
 		return
 	}
 
 	// Use the model's items, type, and required-properties:
-	return referencedDefinition.Properties, c.mapOpenAPITypeToJSONSchemaType(referencedDefinition.Type), referencedDefinition.Required, referencedDefinition.Enum, nil
+	return referencedDefinition.Value.Properties, c.mapOpenAPITypeToJSONSchemaType(referencedDefinition.Value.Type), referencedDefinition.Value.Required, referencedDefinition.Value.Enum, nil
 }
 
 // recurseNestedSchemas converts nested openAPISchemas:
-func (c *Converter) recurseNestedSchemas(nestedSchemas map[string]*openAPI.Schema) (properties map[string]*jsonSchema.Type, err error) {
+func (c *Converter) recurseNestedSchemas(nestedSchemas map[string]*openapi3.SchemaRef) (properties map[string]*jsonSchema.Type, err error) {
 	properties = make(map[string]*jsonSchema.Type)
 
 	// Recurse nested items:
