+++ pkg/schemaconverter/samples/openapi3/array-of-referenced-object.yaml
