+++ pkg/schemaconverter/samples/openapi3/flat-object-with-enum.yaml
