+++ pkg/schemaconverter/samples/openapi3/flat-object-with-number-options.yaml
