+++ pkg/schemaconverter/samples/openapi3/flat-object.yaml
