+++ pkg/schemaconverter/samples/openapi3/object-with-pattern.yaml
