+++ pkg/schemaconverter/samples/openapi3/petstore.yaml
