+++ pkg/schemaconverter/samples/openapi3/referenced-object.yaml
