+++ pkg/schemaconverter/samples/openapi3/with_map.yaml
