+++ pkg/schemaconverter/samples/openapi3/with_map_in_ref_2.yaml
