+++ pkg/schemaconverter/samples/swagger2/array-of-referenced-object.yaml
