+++ pkg/schemaconverter/samples/swagger2/flat-object-with-enum.yaml
