+++ pkg/schemaconverter/samples/swagger2/flat-object-with-number-options.yaml
