+++ pkg/schemaconverter/samples/swagger2/flat-object.yaml
