+++ pkg/schemaconverter/samples/swagger2/object-with-pattern.yaml
