+++ pkg/schemaconverter/samples/swagger2/referenced-object.yaml
