+++ pkg/schemaconverter/samples/swagger2/with_map.yaml
