+++ pkg/schemaconverter/samples/swagger2/with_map_in_ref.yaml
