+++ pkg/schemaconverter/samples/swagger2/with_map_in_ref_2.yaml
