+++ pkg/schemaconverter/types/generated_json_schema.go
