+++ pkg/schemaconverter/types/writer.go
